requirements.txt
faster-whisper>=1.0.3
httpx>=0.27.0
python-dotenv>=1.0.0
zhconv>=1.4.2
yt-dlp>=2025.1.26
